+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Changelog
 All notable changes to this project are documented in this file.
 
+## 2025-11-15
+### Added
+- Frontend Auto-Fix Linting
+### Removed
+- `fetchWorkspaceName` & `workspaceData` as they're no longer used
 
 ## 2025-10-25
+### Added
+- Frontend linting tests
 ### Fixed
 - Appbar Header now removes Header within TodoLists page